project-c/pyproject.toml
[build-system]
requires = ["setuptools >= 68", "wheel","setuptools-git >= 1.1", "pytest-runner", "toml", "reahl-component-metadata >= 7.0.0"]
build-backend = "setuptools.build_meta"

  
[project]
name = "project-c"
version = "4.0"
dependencies = [
  "reahl-component>=7.0,<7.1",
  "reahl-sqlalchemysupport>=7.0,<7.1"
]

[tool.setuptools.packages.find]
exclude = ["etc", "build", "dist"]

[tool.reahl-component]
persisted = [
  "projectc.code:CObject"
]

[tool.reahl-component.versions.'4.0']
migrations = [
  "projectc.migrations:To40"
]

[tool.reahl-component.versions.'3.5']
dependencies = [
  "reahl-component>=7.0,<7.1",
  "reahl-sqlalchemysupport>=7.0,<7.1"
]

migrations = [
  "projectc.migrations:To35"
]

[tool.reahl-component.versions.'3.1']
dependencies = [
  "reahl-component>=7.0,<7.1",
  "reahl-sqlalchemysupport>=7.0,<7.1"
]
migrations = [
  "projectc.migrations:To31"
]

[tool.reahl-component.versions.'3.0']
dependencies = [
  "reahl-component>=7.0,<7.1",
  "reahl-sqlalchemysupport>=7.0,<7.1"
]
migrations = [
  "projectc.migrations:CreateDB"
]

